+++ modernrpc/__init__.py
@@ -1,2 +1,4 @@
+# coding: utf-8
 
+default_app_config = 'modernrpc.apps.ModernRpcConfig'
 __version__ = '0.3.2'