+++ modernrpc/core.py
@@ -2,16 +2,17 @@
 import importlib
 import logging
 import re
+import warnings
 
 from django.core.cache import cache
 from django.core.exceptions import ImproperlyConfigured
 from django.utils import inspect
 
 from modernrpc import modernrpc_settings
-from modernrpc.exceptions import RPCInvalidParams, RPCUnknownMethod, RPCException, RPC_INTERNAL_ERROR
 from modernrpc.handlers import XMLRPC, JSONRPC
 
 logger = logging.getLogger(__name__)
+warnings.simplefilter('once', DeprecationWarning)
 
 RPC_REGISTRY_KEY = '__rpc_registry__'
 DEFAULT_REGISTRY_TIMEOUT = None
@@ -243,6 +244,11 @@ def register_method(function, name=None, entry_point=ALL, protocol=ALL):
     :type protocol: str
     :return: None
     """
+
+    warnings.warn('"register_method" is deprecated and will be removed in a future version.\nUse settings.'
+                  'MODERNRPC_ENTRY_POINTS_MODULES to declare modules containing RPC method, and decorate each method '
+                  'with @rpc_method. Refer\nto the documentation for more info.', DeprecationWarning, stacklevel=2)
+
     # Define the external name of the function
     if not name:
         name = getattr(function, '__name__')
@@ -277,12 +283,54 @@ def register_method(function, name=None, entry_point=ALL, protocol=ALL):
     cache.set(RPC_REGISTRY_KEY, registry, timeout=DEFAULT_REGISTRY_TIMEOUT)
 
 
-def rpc_method(name=None, entry_point=ALL, protocol=ALL):
+def register_rpc_method(function):
+    if not getattr(function, 'modernrpc_enabled', False):
+        raise ImproperlyConfigured('Error: trying to register {} as RPC method, but it has not been decorated.'
+                                   .format(function.__name__))
+
+    # Define the external name of the function
+    name = getattr(function, 'modernrpc_name', function.__name__)
+    logger.debug('Register method {}'.format(name))
+
+    if name.startswith('rpc.'):
+        raise ImproperlyConfigured('According to RPC standard, method names starting with "rpc." are reserved for '
+                                   'system extensions and must not be used. See '
+                                   'http://www.jsonrpc.org/specification#extensions for more information.')
+
+    entry_point = getattr(function, 'modernrpc_entry_point')
+    protocol = getattr(function, 'modernrpc_protocol')
+
+    # Encapsulate the function in a RPCMethod object
+    method = RPCMethod(function, name, entry_point, protocol)
+
+    # Get the current RPC registry from internal cache
+    registry = cache.get(RPC_REGISTRY_KEY, default={})
+
+    # Ensure method names are unique in the registry
+    if method.external_name in registry:
+        # Trying to register many times the same function is OK, because if a method is decorated
+        # with @rpc_method(), it could be imported in different places of the code
+        if method == registry[method.external_name]:
+            return
+        # But if we try to use the same name to register 2 different methods, we
+        # must inform the developer there is an error in the code
+        else:
+            raise ImproperlyConfigured("A RPC method with name {} has already been registered"
+                                       .format(method.external_name))
+
+    # Store the method
+    registry[method.external_name] = method
+    # Update the registry in internal cache
+    cache.set(RPC_REGISTRY_KEY, registry, timeout=DEFAULT_REGISTRY_TIMEOUT)
+
+
+def rpc_method(func=None, name=None, entry_point=ALL, protocol=ALL):
     """
-    Decorator used to define any global function as RPC method.
+    Mark a standard python function as RPC method.
 
     All arguments are optional
 
+    :param func: A standard function
     :param name: Used as RPC method name instead of original function name
     :param entry_point: Default: ALL. Used to limit usage of the RPC method for a specific set of entry points
     :param protocol: Default: ALL. Used to limit usage of the RPC method for a specific protocol (JSONRPC or XMLRPC)
@@ -291,86 +339,19 @@ def rpc_method(name=None, entry_point=ALL, protocol=ALL):
     :type protocol: str
     """
 
-    def __register(function):
-        register_method(function, name, entry_point, protocol)
+    def decorated(function):
+
+        function.modernrpc_enabled = True
+        function.modernrpc_name = name or function.__name__
+        function.modernrpc_entry_point = entry_point
+        function.modernrpc_protocol = protocol
+
         return function
 
-    return __register
-
-
-@rpc_method(name='system.listMethods')
-def __system_listMethods(**kwargs):
-
-    entry_point = kwargs.get(ENTRY_POINT_KEY)
-    protocol = kwargs.get(PROTOCOL_KEY)
-
-    names = [method.name for method in get_all_methods(entry_point, protocol, sort_methods=True)]
-
-    return names
-
-
-@rpc_method(name='system.methodSignature')
-def __system_methodSignature(method_name, **kwargs):
-
-    entry_point = kwargs.get(ENTRY_POINT_KEY)
-    protocol = kwargs.get(PROTOCOL_KEY)
-
-    method = get_method(method_name, entry_point, protocol)
-    if method is None:
-        raise RPCInvalidParams('The method {} is not found in the system. Unable to retrieve signature.')
-    return method.signature
-
-
-@rpc_method(name='system.methodHelp')
-def __system_methodHelp(method_name, **kwargs):
-
-    entry_point = kwargs.get(ENTRY_POINT_KEY)
-    protocol = kwargs.get(PROTOCOL_KEY)
-
-    method = get_method(method_name, entry_point, protocol)
-    if method is None:
-        raise RPCInvalidParams('The method {} is not found in the system. Unable to retrieve method help.')
-    return method.html_doc
-
-
-@rpc_method(name='system.multicall', protocol=XMLRPC)
-def __system_multiCall(calls, **kwargs):
-    """
-    Call multiple RPC methods at once.
-
-    :param calls: An array of struct like {"methodName": string, "params": array }
-    :param kwargs:
-    :return:
-    """
-    if not isinstance(calls, list):
-        raise RPCInvalidParams('method_names must be a list')
-
-    entry_point = kwargs.get(ENTRY_POINT_KEY)
-    protocol = kwargs.get(PROTOCOL_KEY)
-
-    results = []
-    for call in calls:
-        method_name, params = call['methodName'], call['params']
-        method = get_method(method_name, entry_point, protocol)
-
-        try:
-            if not method:
-                raise RPCUnknownMethod(method_name)
-
-            result = method.execute(*params, **kwargs)
-            # From https://mirrors.talideon.com/articles/multicall.html:
-            # "Notice that regular return values are always nested inside a one-element array. This allows you to
-            # return structs from functions without confusing them with faults."
-            results.append([result])
-        except RPCException as e:
-            results.append({
-                'faultCode': e.code,
-                'faultString': e.message,
-            })
-        except Exception as e:
-            results.append({
-                'faultCode': RPC_INTERNAL_ERROR,
-                'faultString': str(e),
-            })
-
-    return results
+    # If @rpc_method is used without any argument nor parenthesis
+    if func is None:
+        def decorator(func):
+            return decorated(func)
+        return decorator
+    # If @rpc_method() is used with parenthesis (with or without arguments)
+    return decorated(func)