+++ modernrpc/tests/testsite/settings.py
@@ -55,3 +55,9 @@ MEDIA_ROOT = ''
 MEDIA_URL = '/'
 STATIC_ROOT = ''
 STATIC_URL = '/'
+
+MODERNRPC_ENTRY_POINTS_MODULES = [
+    'testsite.rpc_methods_stub.generic',
+    'testsite.rpc_methods_stub.specific_types',
+    'testsite.rpc_methods_stub.specific_protocol',
+]